refactor(Pyvium): separate pyvium methods into different classes for better maintainability
